[RFC] Add SPDK as a submodule
This patch starts a series that demonstrates how node driver can utilize SPDK
Storage Management Agent to configure a local SPDK application to connect
remote NVMe/TCP target and expose volumes to be consumed by containers
provisioned by spdk-csi.  Currently, for the sake of simplicity, the SMA and
the SPDK application is assumed to be running on the same host as node driver,
but they could be easily running elsewhere (e.g. on the IPU).

The overall architecture (assuming the use of IPU) looks something like this:

	                   +---------+
	+------------+     |ipu      |
	|node        |     |         |
	|            |  +-->sma---+  |
	|spdk-csi    |  |  |      |  |
	|node driver-+--+  |spdk<-+  |
	|            |     | |       |
	+------------+     +-+-------+
	                     +-----+
	+------------+             |
	|controller  |     +-------+-+
	|            |     |target | |
	|spdk-csi    |     |       | |
	|controller--+----->spdk<--+ |
	|driver      |     |         |
	+------------+     +---------+

The node driver orchestrates the SMA to configure SPDK application to connect
to a target application (previously configured by the controller driver) and
expose volumes, which can then be attached by the node driver to kubernetes
containers.

The submodule points to forked repository that contains the patches adding the
initial version of the Storage Management Agent.  This series is also available
on gerrit [1].

[1] https://review.spdk.io/gerrit/c/spdk/spdk/+/10279

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I353b7b3d8f1d6514a7bb7259163d77ae4ebcfcbc