[RFC] Connecting to a controller during NodeStageVolume
When staging an NVMe/TCP volume, the node driver instructs the SMA to
connect to the controller exposing that volume (and disconnect it when
unstaging the volume).  It doesn't affect the rest of the driver, i.e.
it also connects to the same controller.

The controller IDs are kept on a map with each having its own reference
count.  This refcount stores the number of volumes relying on a
connection to a given controller.  It allows us to disconnect from a
controller once all of its volumes are detached.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I824f4402efd0125d4be72ce978f138cc87488148